requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
appdirs==1.4.4
audiomentations==0.31.0
audioread==3.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.5
cmake==3.25.0
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
docker-pycreds==0.4.0
filelock==3.9.0
fonttools==4.41.0
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
lazy_loader==0.3
librosa==0.10.0.post2
lit==15.0.7
llvmlite==0.40.1
MarkupSafe==2.1.2
matplotlib==3.7.2
ml-insights==1.0.2
mpmath==1.2.1
msgpack==1.0.5
networkx==3.0
numba==0.57.1
numpy==1.24.1
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
Pillow==9.3.0
pooch==1.6.0
protobuf==4.23.4
psutil==5.9.5
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.28.1
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
sentry-sdk==1.28.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
soundfile==0.12.1
soxr==0.3.5
splinecalib==0.0.5
sympy==1.11.1
threadpoolctl==3.2.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wandb==0.15.5
